Moved fil_reader and hdf_reader into their own modules
